https://koliada.atlassian.net/browse/AT-1981
Review changes in 2.0 involving floating point math functions.  Make sure the commands used in the example are the latest versions(say, pow vs fp_pow?).

See the "alternative fp math functions" version of the atmo to see alternative operators that can be used.  That one you can swap between using the two versions of the fp math function calls to ensure they both work and are equivalent.

If you want to replace the fp_cmp function with the >, <, == operators, here's the code to swap in:

float operator_cmp(float a, float b){
    if(a > b){
        return 1;
    }
    else if (a < b){
        return -1;
    }
    else if (a == b){
        return 0;
    }
    return 2;
}