Cloud Outlet Control example.  This will cause the A0-A5 pins to turn on based on the button pressed in the cloud view. Device 1 on = A0 on(high) for 2 seconds. Device 1 off = A1 on(high) for 2 seconds. Device 2 on = A2 on(high) for 2 seconds. Device 2 off = A2 on(high) for 2 seconds. Device 3 on = A3 on(high) for 2 seconds. Device 3 off = A3 on(high) for 2 seconds.
In between these commands being sent, the pins are set to input high impedance mode, so that they won't interfere with someone manually pressing the buttons on their handheld remote.

This is meant to be connected up to a remote control outlet remote which has two buttons for each of the 3 devices(on and off), and operates at around 3 volts.

Because cloud stuff and wifi stuff is broken right now, this has not been fully tested. Functionality needs to be tested before posting once that is possible.